backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "grandalf>=0.8",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.0",
    "langchain-openai>=0.1.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "redis>=5.0.0",
    "tzdata>=2024.1",
    "python-magic-bin>=0.4.14",
]
